+++ src/explicit/explicit_server.cpp
@@ -52,14 +52,14 @@ ExplicitServer::~ExplicitServer()
     UnlinkPointers();
 }
 
-void ExplicitServer::LinkPointers(philote::ExplicitDiscipline *implementation)
+void ExplicitServer::LinkPointers(std::shared_ptr<philote::ExplicitDiscipline> implementation)
 {
     implementation_ = implementation;
 }
 
 void ExplicitServer::UnlinkPointers()
 {
-    implementation_ = nullptr;
+    implementation_.reset();
 }
 
 Status ExplicitServer::ComputeFunction(ServerContext *context,